ENGAGE: Update metadata for refreshed doc
I'd already done the full refresh for this doc in [Work with external groups in Viva Engage networks not aligned to Native Mode](https://review.learn.microsoft.com/en-us/viva/engage/work-with-external-users/create-and-manage-external-groups?branch=main) but missed updating the refresh date. This fixes it.

Original Refresh PR for this is https://github.com/MicrosoftDocs/Viva-pr/pull/9147